requirements.txt
fastapi
uvicorn
pydantic
requests
pdfplumber
python-docx
openai
faiss-cpu
python-dotenv
pymupdf
pinecone